@@ -27,7 +27,9 @@
             <label class="label">Select expense source</label>
             <div class="select">
               <select v-model="expense.note">
-                <option v-for="option in expensees" :key="option.id">{{ option }}</option>
+                <option v-for="option in expensees" :key="option.id">{{
+                  option
+                }}</option>
               </select>
             </div>
           </div>
@@ -38,7 +40,12 @@
               <label class="label has-text-left">Add amount</label>
             </div>
             <div class="field-body">
-              <input class="input" type="number" v-model.trim="expense.value" placeholder="amount">
+              <input
+                class="input"
+                type="number"
+                v-model.trim="expense.value"
+                placeholder="amount"
+              />
             </div>
           </div>
           <div class="field column">
@@ -46,108 +53,125 @@
               <label class="label has-text-left">Select expense date</label>
             </div>
             <div class="field-body">
-              <datepicker input-class="input" placeholder="expense date" v-model="expense.date"></datepicker>
+              <datepicker
+                input-class="input"
+                placeholder="expense date"
+                v-model="expense.date"
+              ></datepicker>
             </div>
           </div>
         </div>
       </section>
       <footer class="modal-card-foot">
-        <button @click="saveExpense" class="button is-success is-rounded">Submit</button>
+        <button @click="saveExpense" class="button is-success is-rounded">
+          Submit
+        </button>
       </footer>
     </div>
   </div>
 </template>
 
 <script>
-  import { mapState } from "vuex";
-  import moment from "moment";
-  import Datepicker from "vuejs-datepicker";
-  const fb = require("../../firebaseConfig");
+import { mapState } from 'vuex'
+import moment from 'moment'
+import Datepicker from 'vuejs-datepicker'
+const fb = require('../../firebaseConfig')
 
-  export default {
-    data() {
-      return {
-        expense: {
-          label: "",
-          category: "",
-          value: null,
-          note: "",
-          date: null
-        }
-      };
-    },
-    components: { Datepicker },
-    computed: {
-      ...mapState(["expenses", "showExpenseModal", "expensees"])
-    },
-    methods: {
-      saveExpense() {
-        let timestamp = moment(this.expense.date).format("MM-DD-YY");
-        let expenseLabel = this.expense.label;
-        let expenseValue = this.expense.value;
-        let billCategory = this.expense.category;
-        let expenseNote = this.expense.note;
-
-        fb.db
-          .collection("Expenses")
-          .add({
-            expense: expenseLabel,
-            category: billCategory,
-            value: expenseValue,
-            note: expenseNote,
-            date: timestamp
-          })
-          .catch(err => alert(err));
-
-        fb.db
-          .collection("ExpenseTotals")
-          .add({
-            date: timestamp,
-            category: expenseLabel,
-            value: expenseValue
-          })
-          .then(() => {
-            (this.expense.label = ""),
-              (this.expense.category = ""),
-              (this.expense.value = null),
-              (this.expense.note = ""),
-              (this.expense.date = "");
-          })
-          .then(() => {
-            this.$store.dispatch("fetchExpenseTotals");
-          });
-
-        fb.db
-          .collection("BudgetedExpenses")
-          .where("name", "==", expenseNote)
-          .get()
-          .then(res => {
-            let doc = res.docs[0].id
-            fb.db
-              .collection("BudgetedExpenses")
-              .doc(doc)
-              .get()
-              .then(d => {
-                  let data = d.data()
-                  let oldSpent = data.spent
-                  
-                  fb.db
-                  .collection("BudgetedExpenses")
-                  .doc(d.id)
-                  .update({
-                      spent: Number(oldSpent) + Number(expenseValue)
-                  })
-              });
-          });
-        this.$modal.hide("expense-modal");
-      },
-      closeModal() {
-        this.$store.state.showExpenseModal = false;
+export default {
+  data() {
+    return {
+      expense: {
+        label: '',
+        category: '',
+        value: null,
+        note: '',
+        date: null
       }
-    },
-    created() {
-      this.$store.dispatch("fetchExpenseTotals");
     }
-  };
-</script>
+  },
+  components: { Datepicker },
+  computed: {
+    ...mapState(['expenses', 'showExpenseModal', 'expensees'])
+  },
+  methods: {
+    saveExpense() {
+      let timestamp = moment(this.expense.date).format('MM-DD-YY')
+      let expenseLabel = this.expense.label
+      let expenseValue = this.expense.value
+      let billCategory = this.expense.category
+      let expenseNote = this.expense.note
+      let uid = this.$store.state.currentUser.uid
 
+      fb.db
+        .collection('Expenses')
+        .doc(uid)
+        .collection('Expenses')
+        .add({
+          expense: expenseLabel,
+          category: billCategory,
+          value: expenseValue,
+          note: expenseNote,
+          date: timestamp
+        })
+        .catch(err => alert(err))
+
+      fb.db
+        .collection('ExpenseTotals')
+        .doc(uid)
+        .collection('Totals')
+        .add({
+          date: timestamp,
+          category: expenseLabel,
+          value: expenseValue
+        })
+        .then(() => {
+          ;(this.expense.label = ''),
+            (this.expense.category = ''),
+            (this.expense.value = null),
+            (this.expense.note = ''),
+            (this.expense.date = '')
+        })
+        .then(() => {
+          this.$store.dispatch('fetchExpenseTotals')
+        })
+
+      fb.db
+        .collection('BudgetedExpenses')
+        .doc(uid)
+        .collection('budgetExpenses')
+        .where('name', '==', expenseNote)
+        .get()
+        .then(res => {
+          let doc = res.docs[0].id
+          fb.db
+            .collection('BudgetedExpenses')
+            .doc(uid)
+            .collection('budgetExpenses')
+            .doc(doc)
+            .get()
+            .then(d => {
+              let data = d.data()
+              let oldSpent = data.spent
+
+              fb.db
+                .collection('BudgetedExpenses')
+                .doc(uid)
+                .collection('budgetExpenses')
+                .doc(d.id)
+                .update({
+                  spent: Number(oldSpent) + Number(expenseValue)
+                })
+            })
+        })
+      this.$modal.hide('expense-modal')
+    },
+    closeModal() {
+      this.$store.state.showExpenseModal = false
+    }
+  },
+  created() {
+    this.$store.dispatch('fetchExpenseTotals')
+    this.$store.dispatch('fetchBaseCategories')
+  }
+}
+</script>
