@@ -11,7 +11,7 @@
                 <Calendar></Calendar>
               </div>
               <div class="column">
-                <ExpensesByDay/>
+                <ExpensesByDay />
               </div>
             </div>
           </div>
@@ -22,34 +22,29 @@
 </template>
 
 <script>
-  import { mapState } from "vuex";
-  import moment from "moment";
-  import numeral from "numeral";
-  import ExpensesByDay from "./charts/ExpensesByDay";
-  import Calendar from "./Calendar";
+import { mapState } from 'vuex'
+import moment from 'moment'
+import numeral from 'numeral'
+import ExpensesByDay from './charts/ExpensesByDay'
+import Calendar from './Calendar'
 
-  export default {
-    components: {
-      ExpensesByDay,
-      Calendar
+export default {
+  components: {
+    ExpensesByDay,
+    Calendar
+  },
+  computed: {
+    ...mapState(['budgetByItems', 'budgetDates', 'expensesByDay'])
+  },
+  filters: {
+    formatDate(val) {
+      let date = moment(val).format('MMM Do, YYYY')
+      return date
     },
-    mounted() {
-      this.$store.dispatch("fetchBudgetItems");
-      this.$store.dispatch("fetchExpensesByDay");
-      this.$store.dispatch("fetchBudgetDates");
-    },
-    computed: {
-      ...mapState(["budgetByItems", "budgetDates", "expensesByDay"]),
-    },
-    filters: {
-      formatDate(val) {
-        let date = moment(val).format("MMM Do, YYYY");
-        return date;
-      },
-      formatCurrency(val) {
-        let value = numeral(val).format("$0,0");
-        return value;
-      }
+    formatCurrency(val) {
+      let value = numeral(val).format('$0,0')
+      return value
     }
-  };
+  }
+}
 </script>
