@@ -1,14 +1,13 @@
 <template>
-    <div>
-        <BillsTable></BillsTable>
-    </div>
+  <div>
+    <BillsTable></BillsTable>
+  </div>
 </template>
 
 <script>
 import BillsTable from './BillsTable'
 
 export default {
-    components: { BillsTable }
+  components: { BillsTable }
 }
 </script>
-
